+++ src/shapes.rs
@@ -1,5 +1,5 @@
 use cgmath::InnerSpace;
-use crate::{Point3, Ray};
+use crate::{Float, Point3, Ray};
 
 pub trait HitObject {
     fn hit(&self, ray: &Ray) -> f64;
@@ -12,17 +12,17 @@ pub enum ShapeEnum {
 #[derive(Clone, Copy, Debug)]
 pub struct Sphere {
     center: Point3,
-    radius: f64,
+    radius: Float,
 }
 
 impl Sphere {
-    pub fn new(center: Point3, radius: f64) -> Sphere {
+    pub fn new(center: Point3, radius: Float) -> Sphere {
         Sphere { center, radius }
     }
 }
 
 impl HitObject for Sphere {
-    fn hit(&self, ray: &Ray) -> f64 {
+    fn hit(&self, ray: &Ray) -> Float {
         let oc = ray.origin - self.center;
         let a = ray.direction.dot(ray.direction);
         let b = 2.0 * oc.dot(ray.direction);
